refactor(FOROME-871): add useScrollShadow hook
