+++ src/components/variant/drawer.utils.ts
@@ -1,3 +1,5 @@
+import { useState } from 'react'
+
 export enum DrawerClass {
   normClass = 'norm',
   normHitClass = 'norm hit',
@@ -12,4 +14,40 @@ export const getLeftDistance = (
   const trackedTdNode = tbodyNode?.children?.[0]?.children?.[1]
 
   return trackedTdNode?.getBoundingClientRect().left
+}
+
+export const useScrollShadow = (
+  element: HTMLDivElement | null,
+): {
+  shouldAddShadow: boolean
+  handleScroll: () => void
+  handleStartScroll: () => void
+} => {
+  const [startedLeftDistance, setStartedLeftDistance] = useState<number | null>(
+    null,
+  )
+
+  const [shouldAddShadow, setShouldAddShadow] = useState(false)
+
+  const handleStartScroll = () => {
+    const currentLeftDistance = getLeftDistance(element)
+
+    if (!currentLeftDistance || startedLeftDistance) return
+
+    const fixedLeftDistance = Math.round(currentLeftDistance)
+    setStartedLeftDistance(fixedLeftDistance)
+  }
+
+  const handleScroll = () => {
+    const currentLeftDistance = getLeftDistance(element)
+
+    if (!currentLeftDistance) return
+
+    const fixedCurrentLeftDistance = Math.round(currentLeftDistance)
+    const isStartPosition = fixedCurrentLeftDistance === startedLeftDistance
+
+    setShouldAddShadow(!isStartPosition)
+  }
+
+  return { shouldAddShadow, handleScroll, handleStartScroll }
 }
